DocumentHandler/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
python-dotenv>=1.0.0

# AI提供商依赖 (可选安装)
requests>=2.28.0
PyJWT>=2.8.0

# 国外AI服务商 (可选)
openai>=1.0.0
anthropic>=0.3.0

# 本地AI支持
# ollama (需要单独安装Ollama服务)

# 开发和测试工具
pytest>=7.0.0
pytest-cov>=4.0.0